refactor(app): move logic into hook
